Add blast http request retry
